fix(types): CI 型チェックの失敗を修正
- build_store_runtime_builder の PEP695 型パラメータに ConfigT を含める
- _proposal_sort_key の返り値注釈を実際の型 (-len は int、metrics は float) に一致させる
- テストのスタブ設定を Protocol を構造的に満たす dataclass に変更
- テストのイベントストア参照に isinstance による型絞り込みを追加

pyright 0 errors / mypy / ty / ruff / pytest 162 passed を確認済み。

+++ src/price_platform/store/keyword_learning/mining.py
@@ -179,7 +179,7 @@ def _select_best_proposal(
     return best_proposal
 
 
-def _proposal_sort_key(proposal: KeywordProposal) -> tuple[float, float, int, int]:
+def _proposal_sort_key(proposal: KeywordProposal) -> tuple[float, float, int, float]:
     return (
         proposal.metrics["precision_lcb"],
         proposal.metrics["estimated_recall_gain"],
